requirements.txt
# Core packages
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1

# RAG components
chromadb==0.4.18
gradio==4.7.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# PyTorch for Apple Silicon (автоматически выберет правильную версию)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0